refactor: move calculateTransactionExpiration to core-utils
